Move internal/x/vars to internal/vars
